Fix FS#8196 - Gather Runtime Data > User Rating not working (don't pass a pointer to short when pointer to int is required)
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@15762 a1c6a512-1295-4272-9138-f99709370657

+++ firmware/export/id3.h
@@ -200,8 +200,8 @@ struct mp3entry {
 
     /* runtime database fields */
     long tagcache_idx;
-    short rating;
-    short score;
+    int rating;
+    int score;
     long playcount;
     long lastplayed;
     long playtime;
